requirements.txt
annotated-types==0.7.0
anyio==4.11.0
apify_client==2.1.0
apify_shared==2.1.0
cachetools==5.5.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
deprecation==2.1.0
google-ai-generativelanguage==0.6.6
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.50.1
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.62.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
impit==0.7.1
joblib==1.5.2
more-itertools==10.8.0
numpy==2.3.3
packaging==24.2
pandas==2.3.3
postgrest==2.21.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.21.1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
storage3==2.21.1
StrEnum==0.4.15
supabase==2.21.1
supabase-auth==2.21.1
supabase-functions==2.21.1
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
yt-dlp==2025.9.26

# Pydantic AI Agent Dependencies
pydantic-ai-slim[openai]==1.18.0
griffe==1.5.1
rich==13.9.4

# Web Dashboard Dependencies
fastapi==0.121.2
uvicorn[standard]==0.32.0
jinja2==3.1.2
slowapi==0.1.9

# Streamlit UI Dependencies
streamlit==1.40.0
